requirements.txt
aiofiles==0.8.0
aioredis==1.3.1
async-timeout==3.0.1
backoff==1.10.0
beautifulsoup4==4.9.3
boto3==1.15.5
botocore==1.18.5
bs4==0.0.1
cachetools==4.2.2
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.9
click==7.1.2
cryptography==3.4.7
dnspython==2.0.0
ecdsa==0.14.1
email-validator==1.1.1
env==0.1.0
Faker==6.6.1
fastapi==0.61.1
fastapi-caching==0.3.0
google-auth==2.3.2
h11==0.9.0
hiredis==2.0.0
idna==2.10
Inject==4.3.1
itsdangerous==2.0.0
jmespath==0.10.0
kubernetes==19.15.0
oauthlib==3.1.1
Pillow==9.2.0
pdf2image==1.16.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.9.0
python-dateutil==2.8.1
python-jose==3.2.0
# Linux
python-magic==0.4.22
# Windows
# python-magic-bin==0.4.14
python-multipart==0.0.5
pytz==2021.1
PyYAML==6.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.3
shortuuid==1.0.1
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
starlette==0.13.6
stringcase==1.2.0
strsimpy==0.2.0
text-unidecode==1.3
typing_extensions==4.2.0
urllib3==1.25.10
uvicorn==0.18.2
websocket-client==1.2.1
websockets==9.1
Werkzeug==1.0.1
stripe==3.5.0
web3==5.30.0
